# Implement fast access mode arrays