--- a/jerry-core/debugger/debugger.c
+++ b/jerry-core/debugger/debugger.c
@@ -15,6 +15,7 @@
 
 #include "byte-code.h"
 #include "debugger.h"
+#include "ecma-array-object.h"
 #include "ecma-builtin-helpers.h"
 #include "ecma-conversion.h"
 #include "ecma-eval.h"
@@ -456,6 +457,17 @@ jerry_debugger_send_scope_variables (const uint8_t *recv_buffer_p) /**< pointer 
   {
     JERRY_ASSERT (ecma_get_lex_env_type (lex_env_p) == ECMA_LEXICAL_ENVIRONMENT_THIS_OBJECT_BOUND);
     ecma_object_t *binding_obj_p = ecma_get_lex_env_binding_object (lex_env_p);
+
+    if (JERRY_UNLIKELY (ecma_get_object_type (binding_obj_p) == ECMA_OBJECT_TYPE_ARRAY))
+    {
+      ecma_extended_object_t *ext_binding_obj_p = (ecma_extended_object_t *) binding_obj_p;
+
+      if (ext_binding_obj_p->u.array.is_fast_mode)
+      {
+        ecma_fast_array_convert_to_normal (binding_obj_p);
+      }
+    }
+
     prop_iter_cp = binding_obj_p->u1.property_list_cp;
   }