--- a/jerry-core/ecma/builtin-objects/ecma-builtins.c
+++ b/jerry-core/ecma/builtin-objects/ecma-builtins.c
@@ -432,6 +432,8 @@ ecma_instantiate_builtin (ecma_builtin_id_t obj_builtin_id) /**< built-in id */
 
       ext_object_p->u.array.length = 0;
       ext_object_p->u.array.length_prop = ECMA_PROPERTY_FLAG_WRITABLE | ECMA_PROPERTY_TYPE_VIRTUAL;
+      ext_object_p->u.array.is_fast_mode = false;
+      ext_object_p->u.array.hole_count = 0;
       break;
     }
 #endif /* ENABLED (JERRY_BUILTIN_ARRAY) */