… jerry-core/ecma/base/ecma-helpers.c
… jerry-core/ecma/operations/ecma-array-object.c