@rerobika 2019-08-28
Benchmark | Heap (bytes) | RSS (bytes) | Perf (sec) |
----: | ----: | ----: | ----: | 
3d-cube.js | 16376 -> 16376 : 0.000% | 90112 -> 90112 : 0.000% | 0.764 -> 0.693 : +9.250% | 
3d-raytrace.js | 121872 -> 110312 : +9.485% | 172617 -> 154477 : +10.509% | 0.908 -> 0.867 : +4.528% | 
access-binary-trees.js | 32752 -> 32752 : 0.000% | 98304 -> 98304 : 0.000% | 0.452 -> 0.453 : -0.251% | 
access-fannkuch.js | 4960 -> 4840 : +2.419% | 69632 -> 69632 : 0.000% | 1.777 -> 1.660 : +6.590% | 
access-nbody.js | 6264 -> 6200 : +1.022% | 77824 -> 77824 : 0.000% | 0.990 -> 0.979 : +1.126% | 
bitops-3bit-bits-in-byte.js | 712 -> 712 : 0.000% | 65536 -> 65536 : 0.000% | 0.465 -> 0.467 : -0.278% | 
bitops-bits-in-byte.js | 664 -> 664 : 0.000% | 65536 -> 65536 : 0.000% | 0.628 -> 0.628 : -0.000% | 
bitops-bitwise-and.js | 496 -> 496 : 0.000% | 65536 -> 65536 : 0.000% | 0.769 -> 0.770 : -0.117% | 
bitops-nsieve-bits.js | 98304 -> 60880 : +38.070% | 163840 -> 126976 : +22.500% | 1.079 -> 0.971 : +9.958% | 
controlflow-recursive.js | 888 -> 888 : 0.000% | 65536 -> 65536 : 0.000% | 0.320 -> 0.327 : -1.945% | 
crypto-aes.js | 33200 -> 32752 : +1.349% | 114688 -> 114688 : 0.000% | 0.771 -> 0.591 : +23.341% | 
crypto-md5.js | 106488 -> 73712 : +30.779% | 180224 -> 159744 : +11.364% | 0.511 -> 0.500 : +2.047% | 
crypto-sha1.js | 65528 -> 49144 : +25.003% | 135168 -> 126976 : +6.061% | 0.515 -> 0.503 : +2.441% | 
date-format-tofte.js | 16368 -> 8184 : +50.000% | 90112 -> 81920 : +9.091% | 0.697 -> 0.679 : +2.523% | 
date-format-xparb.js | 16368 -> 16368 : 0.000% | 90112 -> 90112 : 0.000% | 0.485 -> 0.490 : -1.030% | 
math-cordic.js | 1728 -> 1696 : +1.852% | 69632 -> 69632 : 0.000% | 1.063 -> 1.056 : +0.706% | 
math-partial-sums.js | 1240 -> 1240 : 0.000% | 65536 -> 65536 : 0.000% | 0.640 -> 0.628 : +1.834% | 
math-spectral-norm.js | 6096 -> 4632 : +24.016% | 73728 -> 69632 : +5.556% | 0.461 -> 0.457 : +0.733% | 
string-base64.js | 90800 -> 89728 : +1.181% | 192512 -> 188416 : +2.128% | 1.221 -> 1.196 : +2.083% | 
string-fasta.js | 8256 -> 8256 : 0.000% | 73728 -> 73728 : 0.000% | 1.189 -> 1.217 : -2.376% | 
Geometric mean: | +10.779% | +3.551% | +3.242% | 

Now we are able to run 3d-morph.js:
Before the PR:
 - 522KB memory
 - 3900ms runtime

After the PR:
 - 289KB mermory
 - 87ms runtime

Binary (bytes) | master(47f2f0ea8b) | patch(5e2a66f324) | Diff |
----: | ----: | ----: | ----: |
size | 128568 | 132664 | +4096 bytes |
.rodata | 11722 | 11730 | +8 bytes |
.dynstr | 262 | 262 | 0 bytes |
.rel.plt | 176 | 176 | 0 bytes |
.interp | 25 | 25 | 0 bytes |
.dynsym | 384 | 384 | 0 bytes |
.gnu.hash | 188 | 188 | 0 bytes |
.text | 111860 | 113516 | +1656 bytes |
.comment | 59 | 59 | 0 bytes |
.shstrtab | 250 | 250 | 0 bytes |
.data | 8 | 8 | 0 bytes |
.ARM.exidx | 8 | 8 | 0 bytes |
.rel.dyn | 16 | 16 | 0 bytes |
.init | 12 | 12 | 0 bytes |
.got | 104 | 104 | 0 bytes |
.plt | 288 | 288 | 0 bytes |
.note.gnu.build-id | 36 | 36 | 0 bytes |
.note.ABI-tag | 32 | 32 | 0 bytes |
.gnu.version_r | 64 | 64 | 0 bytes |
.bss | 1575096 | 1575096 | 0 bytes |
.fini | 8 | 8 | 0 bytes |
.gnu.version | 48 | 48 | 0 bytes |
.fini_array | 4 | 4 | 0 bytes |
.init_array | 4 | 4 | 0 bytes |
.dynamic | 232 | 232 | 0 bytes |
.eh_frame | 4 | 4 | 0 bytes |
.ARM.attributes | 51 | 51 | 0 bytes |
.jcr | 4 | 4 | 0 bytes |